HW-52174. Moved date fragment outside of widget
